+++ ft_memccpy.c
@@ -6,11 +6,13 @@
 /*   By: trbonnes <trbonnes@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/08 08:23:10 by trbonnes          #+#    #+#             */
-/*   Updated: 2019/10/08 12:17:20 by trbonnes         ###   ########.fr       */
+/*   Updated: 2019/10/08 16:32:52 by trbonnes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memccpy(void *dest, const void *src, int c, unsigned int n)
+#include <stdlib.h>
+
+void	*ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
 	unsigned char	*ptrd;
 	const char		*ptrs;
