+++ ft_strncmp.c
@@ -6,13 +6,15 @@
 /*   By: trbonnes <trbonnes@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/08 13:29:36 by trbonnes          #+#    #+#             */
-/*   Updated: 2019/10/08 13:51:50 by trbonnes         ###   ########.fr       */
+/*   Updated: 2019/10/08 16:54:02 by trbonnes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+#include <stdlib.h>
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int i;
+	size_t i;
 
 	i = 0;
 	while (i < n)
